Add files via upload
 STEGANOGRAPHY PROJECT

The Steganography project in C is a console-based application that demonstrates how to hide (encode) a secret message or file inside a cover file (like a BMP image or a text file) without visibly altering the original file. It also supports decoding the hidden data. The project focuses on understanding data hiding, bitwise operations, and file handling using the C programming language.